feat: combine power and workout controls
Since they both are small, I think it is easiest to just combine them.
Current solution makes it a bit hard to remember which one to click and such.
Could ofc be solved another way (different icons). But I think this is an easier solution both for us and users.

One could then delete one of the buttons, but I just kept both, since it is so symmetric :))